Refactored class tree for streaming validations.
