+++ build.sbt
@@ -295,7 +295,7 @@ lazy val cometVersion  = "0.1.7"
 lazy val mongodbVersion      = "4.6.0"
 lazy val mongo4catsVersion   = "0.4.7"
 lazy val any23Version        = "2.7"
-lazy val rdf4jVersion        = "3.7.6"
+lazy val rdf4jVersion        = "4.0.0"
 lazy val graphvizJavaVersion = "0.18.1"
 lazy val logbackVersion      = "1.2.11"
 lazy val scalaLoggingVersion = "3.9.4"
