+++ modules/server/src/main/scala/es/weso/rdfshape/server/api/routes/schema/logic/operations/stream/configuration/StreamValidationExtractorConfiguration.scala
@@ -1,4 +1,4 @@
-package es.weso.rdfshape.server.api.routes.schema.service.operations.stream.configuration
+package es.weso.rdfshape.server.api.routes.schema.logic.operations.stream.configuration
 
 import cats.implicits.toBifunctorOps
 import es.weso.rdf.InferenceEngine
