+++ modules/server/src/main/scala/es/weso/rdfshape/server/api/routes/schema/logic/operations/stream/transformations/ValidationResultTransformations.scala
@@ -1,4 +1,4 @@
-package es.weso.rdfshape.server.api.routes.schema.logic.stream.transformations
+package es.weso.rdfshape.server.api.routes.schema.logic.operations.stream.transformations
 
 import es.weso.rdfshape.server.api.utils.parameters.IncomingRequestParameters.{
   ContentParameter,
